Rename NotificationBanner into NotificationCard
