+++ packages/circuit-ui/components/Notification/Notification.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../../.storybook/components';
-import NotificationBanner from '../NotificationBanner';
+import NotificationCard from '../NotificationCard';
 import NotificationList from '../NotificationList';
 
 # Notification
@@ -41,12 +41,12 @@ to use a call to action.
 
 **Error**: When action is required by the user, attention notifications should be used and should clearly highlight which action is required. **A call to action must be displayed.**
 
-## NotificationBanner
+## NotificationCard
 
-You can display a banner with a notification with the `NotificationBanner`, and position it at the top or bottom of a view.
+You can display a banner with a notification with the `NotificationCard`, and position it at the top or bottom of a view.
 
 <Story id="components-notification-notificationbanner--base" />
-<Props of={NotificationBanner} />
+<Props of={NotificationCard} />
 
 ## NotificationList
